% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_model}
\alias{check_model}
\title{Checks whether the input \code{model} is valid.}
\usage{
check_model(model)
}
\arguments{
\item{model}{A \code{model} object. See \link{example_model} for details.}
}
\description{
Checks whether the input \code{model} is valid.
}
