% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{example_clustering}
\alias{example_clustering}
\title{An example \code{clustering} object. A clustering is a \code{List} that
splits indexes 1..#num_datapoints to clusters. Each \code{List} element corresponds to one cluster.
The clustering is not necessarily a partition but it usually is.}
\usage{
example_clustering()
}
\value{
A \code{List} for the clustering of indexes 1..#num_datapoints.
}
\description{
An example \code{clustering} object. A clustering is a \code{List} that
splits indexes 1..#num_datapoints to clusters. Each \code{List} element corresponds to one cluster.
The clustering is not necessarily a partition but it usually is.
}
