% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastLm}
\alias{fastLm}
\title{Fast least squares}
\usage{
fastLm(y, X)
}
\arguments{
\item{y}{Vector of outcomes.}

\item{X}{Matrix of covariates (first column should be 1's)}
}
\value{
\code{List} of regression output with elements \code{coef}, \code{stderr}.
}
\description{
This functions fits the regression y ~ X using Armadillo \code{solve} function.
}
