% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OLS_c}
\alias{OLS_c}
\title{Fast least squares}
\usage{
OLS_c(y, X)
}
\arguments{
\item{y}{Vector of outcomes.}

\item{X}{Matrix of covariates (first column should be 1's)}
}
\value{
Vector of coefficients.
}
\description{
Fast OLS as in \link{fastLm} but returns only the fitted coefficients.
}
