% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{out_pval}
\alias{out_pval}
\title{Calculates p-value or test decision}
\usage{
out_pval(rtest_out, ret_pval, alpha)
}
\arguments{
\item{rtest_out}{A \code{List} with elements \code{tobs}, \code{tvals} (see \link{one_sided_test} for details.)}

\item{ret_pval}{A \code{Boolean} indicating whether to return a p-value (TRUE) or not.}

\item{alpha}{Desired test level (from 0 to 1).}
}
\value{
Binary decision if \code{ret_pval} is TRUE, or the p-value otherwise.
}
\description{
Depending on \code{ret_pval} this function returns either a p-value for the test or the binary decision.
}
\details{
Returns 1 if the test rejects, 0 otherwise.
}
